Bump Apollo Server to 3 and supported graphql-tool packages
