+++ services/server/package.json
@@ -25,11 +25,13 @@
     "typecheck": "tsc -b tsconfig.app.json tsconfig.crons.json tsconfig.libs.json"
   },
   "dependencies": {
-    "@graphql-tools/load-files": "6.0.8",
-    "@graphql-tools/merge": "6.0.8",
+    "@graphql-tools/load-files": "6.3.2",
+    "@graphql-tools/merge": "6.2.14",
+    "@graphql-tools/schema": "7.1.5",
+    "@graphql-tools/utils": "7.10.0",
     "@route-codegen/utils": "0.2.0",
-    "apollo-server-express": "2.25.1",
-    "apollo-server-plugin-base": "0.13.0",
+    "apollo-server-express": "3.0.1",
+    "apollo-server-plugin-base": "3.0.1",
     "bcryptjs": "2.4.3",
     "cookie": "0.4.0",
     "cookie-parser": "1.4.5",
