+++ services/server/src/app/graph/directiveResolvers/directiveResolvers.ts
@@ -1,0 +1,8 @@
+import { DirectiveResolvers } from "@libs/graph/resolvers.generated";
+import { isLoggedIn } from "./isLoggedIn";
+import { isPrivate } from "./isPrivate";
+
+export const directiveResolvers: DirectiveResolvers = {
+  isLoggedIn,
+  isPrivate,
+};
