--- services/server/src/app/graph/directives/index.ts
@@ -1,2 +1,0 @@
-export * from "./IsLoggedInDirective";
-export * from "./IsPrivateDirective";
