+++ services/server/src/app/web/plugins/createErrorNotifier.ts
@@ -2,9 +2,9 @@ import { ApolloServerPlugin } from "apollo-server-plugin-base";
 
 export const createErrorNotifier = (): ApolloServerPlugin => {
   return {
-    requestDidStart() {
+    async requestDidStart() {
       return {
-        didEncounterErrors(requestContext) {
+        async didEncounterErrors(requestContext) {
           const [error] = requestContext.errors;
           if (error) {
             console.error(error);
