+++ services/server/src/app/web/server.ts
@@ -18,7 +18,7 @@ const jwt = createJwtService({
   jwtSecret: env.JWT_SECRET,
 });
 
-const { httpServer, apolloServer } = createServers({
+createServers({
   stage: STAGE,
   corsOptions: {
     origin: [`https://${env.CLIENT_SERVICE_DOMAIN}`, `https://${env.CLIENT_SEO_SERVICE_DOMAIN}`],
@@ -30,9 +30,8 @@ const { httpServer, apolloServer } = createServers({
     headersService: headers,
     jwtService: jwt,
   },
-});
-
-httpServer.listen(PORT, () => {
-  console.log(`🚀 Server ready at https://localhost:${PORT}${apolloServer.graphqlPath}`);
-  console.log(`🚀 Subscriptions ready at wss://localhost:${PORT}${apolloServer.subscriptionsPath}`);
+}).then(({ apolloServer, expressServer }) => {
+  expressServer.listen({ port: PORT }, () => {
+    console.log(`🚀 Server ready at http://localhost:${PORT}${apolloServer.graphqlPath}`);
+  });
 });
