+++ services/server/tsconfig.base.json
@@ -3,6 +3,7 @@
     "lib": ["es2016", "esnext.asynciterable"],
     "target": "es2016",
     "module": "commonjs",
+    "moduleResolution": "node",
     "sourceMap": true,
     "strict": true,
     "skipLibCheck": true,
