+++ services/server/tsconfig.json
@@ -2,6 +2,7 @@
   "compilerOptions": {
     "target": "es2016",
     "module": "commonjs",
+    "moduleResolution": "node",
     "sourceMap": true,
     "strict": true,
     "skipLibCheck": true,
